+++ tiledb/query_condition.py
@@ -21,7 +21,18 @@ QueryConditionNodeElem = Union[
 class QueryCondition:
     """
     Class representing a TileDB query condition object for attribute filtering
-    pushdown. Set the query condition with a string representing an expression
+    pushdown.
+
+    When querying a sparse array, only the values that satisfy the given
+    condition are returned (coupled with their associated coordinates). An example
+    may be found in `examples/query_condition_sparse.py`.
+
+    For dense arrays, the given shape of the query matches the shape of the output
+    array. Values that DO NOT satisfy the given condition are filled with the
+    TileDB default fill value. Different attribute types have different default
+    fill values as outlined here (https://docs.tiledb.com/main/background/internal-mechanics/writing#default-fill-values). An example may be found in `examples/query_condition_dense.py`.
+
+    Set the query condition with a string representing an expression
     as defined by the grammar below. A more straight forward example of usage is
     given beneath.